requirements.txt
python==3.9.7
asgiref==3.4.1
certifi==2020.6.20
charset-normalizer==2.0.10
Django==4.0.1
djangorestframework==3.13.1
idna==3.3
joblib==1.1.0
lxml==4.7.1
mysqlclient==2.1.0
numpy==1.22.1
pandas==1.3.5
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tzdata==2021.5
urllib3==1.26.8
wincertstore==0.2
